backend/requirements.txt
fastapi
uvicorn
pydantic
langchain
openai
chromadb
kafka-python